Setup env for API token #GSM-22 Rename util to utilities & add env variables to apiUtil file
