--- frontend/constants/apiEndpoints.js
@@ -1,3 +1,0 @@
-export const iexApiFreeUrl = 'https://api.iextrading.com/1.0';
-export const iexApiCloudUrl = 'https://cloud.iexapis.com';
-export const iexApiSandboxUrl = 'https://sandbox.iexapis.com';
